Fixed camera zoom

+++ src/main/kotlin/com/cout970/modeler/render/controller/Camera.kt
@@ -24,7 +24,7 @@ data class Camera(
 
     val matrix by lazy {
         Matrix4d().apply {
-            translate(0.0, 0.0, -zoom.toDouble())
+            translate(0.0, 0.0, -zoom)
             rotate(angleY, 1.0, 0.0, 0.0)
             rotate(angleX, 0.0, 1.0, 0.0)
             translate(position.xd, position.yd, position.zd)
